Used radar grid blocking, moved Modulate functions into their own namespace
