+++ cxx/isce3/Headers.cmake
@@ -114,6 +114,7 @@ geometry/metadataCubes.h
 geogrid/getRadarGrid.h
 geogrid/relocateRaster.h
 image/forward.h
+image/Modulate.h
 image/Resample.h
 image/ResampSlc.h
 image/ResampSlc.icc
