+++ cxx/isce3/Sources.cmake
@@ -59,6 +59,7 @@ geometry/TopoLayers.cpp
 geometry/metadataCubes.cpp
 geogrid/getRadarGrid.cpp
 geogrid/relocateRaster.cpp
+image/Modulate.cpp
 image/Resample.cpp
 image/ResampSlc.cpp
 io/gdal/Dataset.cpp
