+++ cxx/isce3/image/Resample.cpp
@@ -7,203 +7,6 @@
 
 namespace isce3::image::v2 {
 
-
-template<typename AzRgFunc = isce3::core::Poly2d>
-std::complex<double> _getPixelCarrierPhase(
-    const double azimuth,
-    const double range,
-    const AzRgFunc& carrier_phase,
-    bool conjugate = false
-) {
-      // Evaluate the pixel's carrier phase
-      // unit: phase (radians)
-      const double phase = carrier_phase.eval(azimuth, range);
-
-      // Convert the carrier phase into a unit phasor (i.e. an angle on the unit
-      // circle in the complex plane).
-      // unit: unitless (complex)
-      if(conjugate){
-          return std::complex<double>(std::cos(phase), -std::sin(phase));
-      } else {
-          return std::complex<double>(std::cos(phase), std::sin(phase));
-      }
-}
-
-
-template<typename AzRgFunc>
-void getModulationPhase(
-    ArrayRef2D<std::complex<float>> out,
-    const AzRgFunc& carrier_phase,
-    const isce3::product::RadarGridParameters& radar_grid,
-    const size_t input_azimuth_first_line,
-    const size_t input_range_first_pixel,
-    const bool conjugate
-)
-{
-    // unit: azimuth row indices (int)
-    const size_t phase_block_length = out.rows();
-    // unit: range column indices (int)
-    const size_t phase_block_width = out.cols();
-
-    // remove carrier from radar data
-#pragma omp parallel for collapse(2)
-    for (size_t az_index = 0; az_index < phase_block_length; ++az_index) {
-        for (size_t rg_index = 0; rg_index < phase_block_width; ++rg_index) {
-            // Offset for block starting line
-            // unit: azimuth row indices (int)
-            const auto az_carrier_index = az_index + input_azimuth_first_line;
-            // unit: time (seconds)
-            const double azimuth =
-                radar_grid.sensingStart() + az_carrier_index / radar_grid.prf();
-
-            // Offset for block starting pixel
-            // unit: range column indices (int)
-            const auto rg_carrier_index = rg_index + input_range_first_pixel;
-            // unit: distance (meters)
-            const double range = radar_grid.startingRange() +
-                    rg_carrier_index * radar_grid.rangePixelSpacing();
-            
-            // Get the carrier unit phasor for this pixel
-            const auto phasor = _getPixelCarrierPhase(
-                azimuth, range, carrier_phase, conjugate
-            );
-            
-            // Write the phasor into the output data block
-            out(az_index, rg_index) = phasor;
-        }
-    } // end multithreaded block
-}
-
-
-template<typename AzRgFunc>
-void getModulationPhaseAtCoords(
-    ArrayRef2D<std::complex<float>> out,
-    const AzRgFunc& carrier_phase,
-    const isce3::product::RadarGridParameters& radar_grid,
-    const ConstArrayRef2D<double> azimuth_indices,
-    const ConstArrayRef2D<double> range_indices,
-    const bool conjugate
-)
-{
-    const size_t outWidth = out.cols();
-    const size_t outLength = out.rows();
-
-#pragma omp parallel for collapse(2)
-    for (size_t az_index = 0; az_index < outLength; ++az_index){
-        for (size_t rg_index = 0; rg_index < outWidth; ++rg_index){
-
-            // Get the indices on the output grid corresponding to these indices
-            // on the input grid overall.
-            const double az_carrier_index = azimuth_indices(az_index, rg_index);
-            const double rg_carrier_index = range_indices(az_index, rg_index);
-
-            // Azimuth time at the current output pixel
-            const double azimuth = radar_grid.sensingStart() + az_carrier_index / 
-                radar_grid.prf();
-
-            // Slant Range at the current output pixel
-            const double range = radar_grid.startingRange() + rg_carrier_index *
-                radar_grid.rangePixelSpacing();
-            
-            // Get the carrier phasor for this pixel
-            const auto phasor = _getPixelCarrierPhase(
-                azimuth, range, carrier_phase, conjugate
-            );
-            
-            // Write the phasor into the output data block
-            out(az_index, rg_index) = phasor;
-        }
-    } // end multithreaded block
-}
-
-
-template<typename AzRgFunc>
-void modulate(
-    ArrayRef2D<std::complex<float>> slc_data_block,
-    const AzRgFunc& carrier_phase,
-    const isce3::product::RadarGridParameters& radar_grid,
-    const size_t input_azimuth_first_line,
-    const size_t input_range_first_pixel,
-    const bool conjugate
-)
-{
-    // unit: azimuth row indices (int)
-    const size_t phase_block_length = slc_data_block.rows();
-    // unit: range column indices (int)
-    const size_t phase_block_width = slc_data_block.cols();
-
-    // remove carrier from radar data
-#pragma omp parallel for collapse(2)
-    for ( size_t az_index = 0; az_index < phase_block_length; ++az_index) {
-        for (size_t rg_index = 0; rg_index < phase_block_width; ++rg_index) {
-            // Offset for block starting line
-            // unit: azimuth row indices (int)
-            const auto az_carrier_index = az_index + input_azimuth_first_line;
-            // unit: time (seconds)
-            const double azimuth =
-                radar_grid.sensingStart() + az_carrier_index / radar_grid.prf();
-
-            // Offset for block starting pixel
-            // unit: range column indices (int)
-            const auto rg_carrier_index = rg_index + input_range_first_pixel;
-            // unit: distance (meters)
-            const double range = radar_grid.startingRange() +
-                    rg_carrier_index * radar_grid.rangePixelSpacing();
-            
-            // Get the carrier phasor for this pixel
-            const auto phasor = _getPixelCarrierPhase(
-                azimuth, range, carrier_phase, conjugate
-            );
-            
-            // Modulate the phasor into the output data block
-            slc_data_block(az_index, rg_index) *= phasor;
-        }
-    } // end multithreaded block
-}
-
-
-template<typename AzRgFunc>
-void modulateAtCoords(
-    ArrayRef2D<std::complex<float>> slc_data_block,
-    const AzRgFunc& carrier_phase,
-    const isce3::product::RadarGridParameters& radar_grid,
-    const ConstArrayRef2D<double> azimuth_indices,
-    const ConstArrayRef2D<double> range_indices,
-    const bool conjugate
-)
-{
-    const size_t outWidth = slc_data_block.cols();
-    const size_t outLength = slc_data_block.rows();
-
-#pragma omp parallel for collapse(2)
-    for (size_t az_index = 0; az_index < outLength; ++az_index){
-        for (size_t rg_index = 0; rg_index < outWidth; ++rg_index){
-
-            // Get the indices on the output grid corresponding to these indices
-            // on the input grid overall.
-            const double az_carrier_index = azimuth_indices(az_index, rg_index);
-            const double rg_carrier_index = range_indices(az_index, rg_index);
-
-            // Azimuth time at the current output pixel
-            const double azimuth = radar_grid.sensingStart() + az_carrier_index / 
-                radar_grid.prf();
-
-            // Slant Range at the current output pixel
-            const double range = radar_grid.startingRange() + rg_carrier_index *
-                radar_grid.rangePixelSpacing();
-            
-            // Get the carrier phasor for this pixel
-            const auto phasor = _getPixelCarrierPhase(
-                azimuth, range, carrier_phase, conjugate
-            );
-            
-            // Modulate the phasor into the output data block
-            slc_data_block(az_index, rg_index) *= phasor;
-        }
-    } // end multithreaded block
-}
-
-
 void resampleToCoords(
     ArrayRef2D<std::complex<float>> resampled_data_block,
     const ConstArrayRef2D<std::complex<float>> input_data_block,
@@ -370,42 +173,5 @@ void resampleToCoords(
         } // end omp for collapse(2)
     } // end omp parallel
 } // end resampleToCoords
-
-
-#define EXPLICIT_INSTANTIATION(AzRgFunc)                                      \
-template void getModulationPhase(                                             \
-    ArrayRef2D<std::complex<float>> out,                                      \
-    const AzRgFunc& carrier_phase,                                            \
-    const isce3::product::RadarGridParameters& radar_grid,                    \
-    const size_t input_azimuth_first_line,                                    \
-    const size_t input_range_first_pixel,                                     \
-    const bool conjugate                                                      \
-);                                                                            \
-template void getModulationPhaseAtCoords(                                     \
-    ArrayRef2D<std::complex<float>> out,                                      \
-    const AzRgFunc& carrier_phase,                                            \
-    const isce3::product::RadarGridParameters& radar_grid,                    \
-    const ConstArrayRef2D<double> azimuth_indices,                            \
-    const ConstArrayRef2D<double> range_indices,                              \
-    const bool conjugate                                                      \
-);                                                                            \
-template void modulate(                                                       \
-    ArrayRef2D<std::complex<float>> slc_data_block,                           \
-    const AzRgFunc& carrier_phase,                                            \
-    const isce3::product::RadarGridParameters& radar_grid,                    \
-    const size_t input_azimuth_first_line,                                    \
-    const size_t input_range_first_pixel,                                     \
-    const bool conjugate                                                      \
-);                                                                            \
-template void modulateAtCoords(                                               \
-    ArrayRef2D<std::complex<float>> slc_data_block,                           \
-    const AzRgFunc& carrier_phase,                                            \
-    const isce3::product::RadarGridParameters& radar_grid,                    \
-    const ConstArrayRef2D<double> azimuth_indices,                            \
-    const ConstArrayRef2D<double> range_indices,                              \
-    const bool conjugate                                                      \
-)
-EXPLICIT_INSTANTIATION(isce3::core::LUT2d<double>);
-EXPLICIT_INSTANTIATION(isce3::core::Poly2d);
 
 } // end namespace isce3::image::v2
