+++ python/extensions/pybind_isce3/Sources.cmake
@@ -59,6 +59,7 @@ geogrid/relocateRaster.cpp
 geogrid/geogrid.cpp
 geometry/lookIncFromSr.cpp
 image/image.cpp
+image/Modulate.cpp
 image/Resample.cpp
 image/ResampSlc.cpp
 io/gdal/Dataset.cpp
