+++ python/extensions/pybind_isce3/image/Modulate.h
@@ -1,0 +1,5 @@
+#pragma once
+#include <pybind11/pybind11.h>
+
+template<typename T>
+void addbindings_modulate(pybind11::module&);
